changed tailwind.css
