+++ auto-connect-frontend/src/components/NavBar.jsx
@@ -1,188 +1,129 @@
-import React, { useState, useEffect } from "react";
-import { Link } from "react-router-dom";
-import { motion, AnimatePresence } from "framer-motion";
-import "./Navbar.css";
+import { Link, useLocation, matchPath } from "react-router-dom";
+import "./NavBar.css";
 
-const Navbar = ({ onSignIn }) => {
-  const [isScrolled, setIsScrolled] = useState(false);
-  const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
+function NavBar({ navLinks, version, icon, userId }) {
+  const location = useLocation();
+  const path = location.pathname;
 
-  useEffect(() => {
-    const handleScroll = () => {
-      setIsScrolled(window.scrollY > 50);
-    };
+  const isActive = (routePaths) =>
+    Array.isArray(routePaths)
+      ? routePaths.some((p) => matchPath({ path: p, end: true }, path))
+      : false;
 
-    window.addEventListener("scroll", handleScroll);
-    return () => window.removeEventListener("scroll", handleScroll);
-  }, []);
+  const isNavLinkActive = (link) => {
+    console.log("Checking if link is active:", link);
+    if (link.path && isActive(link.path)) {
+      return true;
+    }
 
-  const navLinks = [
-    { name: "Features", href: "#features" },
-    { name: "How It Works", href: "#how-it-works" },
-    { name: "Testimonials", href: "#testimonials" },
-    { name: "Contact", href: "#contact" },
-  ];
+    if (link.sub) {
+      if (isActive(link.sub.path)) {
+        return true;
+      }
 
-  const handleNavClick = (href) => {
-    if (href.startsWith("#")) {
-      const element = document.querySelector(href);
-      if (element) {
-        element.scrollIntoView({ behavior: "smooth" });
+      if (link.subNav) {
+        return true;
       }
     }
-    setIsMobileMenuOpen(false);
+
+    if (link.subNav) {
+      return link.subNav.some((subLink) => isActive(subLink.path));
+    }
+
+    return false;
   };
 
-  return (
-    <motion.nav
-      initial={{ y: -100 }}
-      animate={{ y: 0 }}
-      transition={{ duration: 0.6, ease: "easeOut" }}
-      className={`navbar ${isScrolled ? "scrolled" : ""}`}
-    >
-      <div className="nav-container">
-        {/* Logo */}
-        <motion.div whileHover={{ scale: 1.05 }} className="nav-logo">
-          <Link to="/" className="logo-link">
-            <div className="logo-icon">
-              <svg
-                viewBox="0 0 24 24"
-                fill="none"
-                xmlns="http://www.w3.org/2000/svg"
-              >
-                <path
-                  d="M3 18V16C3 14.8954 3.89543 14 5 14H19C20.1046 14 21 14.8954 21 16V18M3 18V20C3 21.1046 3.89543 22 5 22H19C20.1046 22 21 21.1046 21 20V18M3 18H21M12 2L8 6H16L12 2Z"
-                  stroke="currentColor"
-                  strokeWidth="2"
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                />
-                <circle cx="7" cy="18" r="1" fill="currentColor" />
-                <circle cx="17" cy="18" r="1" fill="currentColor" />
-              </svg>
-            </div>
-            <span className="logo-text">AutoConnect</span>
-          </Link>
-        </motion.div>
+  if (!version) {
+    version = "0.0.0v";
+  }
 
-        {/* Desktop Navigation */}
-        <div className="nav-links desktop-nav">
-          {navLinks.map((link, index) => (
-            <motion.a
-              key={link.name}
-              href={link.href}
-              onClick={(e) => {
-                e.preventDefault();
-                handleNavClick(link.href);
-              }}
-              className="nav-link"
-              whileHover={{ y: -2 }}
-              initial={{ opacity: 0, y: -20 }}
-              animate={{ opacity: 1, y: 0 }}
-              transition={{ delay: index * 0.1 + 0.3 }}
+  if (!icon) {
+    icon = "";
+  }
+
+  const navLinkElements = navLinks.map((link, index) => {
+    const ContainerType = link.sub ? "span" : Link;
+
+    return (
+      <li
+        key={index}
+        className={`nav-link ${isNavLinkActive(link) ? "selected" : ""}`}
+        title={link.title}
+      >
+        <label htmlFor={index} className="link-label">
+          <span className="nav-link__main">
+            <ContainerType
+              to={link.path ? link.path[0] : ""}
+              className={`nav-link__details nav-link ${
+                isNavLinkActive(link) ? "selected" : ""
+              }`}
             >
-              {link.name}
-            </motion.a>
-          ))}
-        </div>
-
-        {/* Auth Buttons */}
-        <div className="nav-auth desktop-nav">
-          <motion.button
-            onClick={onSignIn}
-            className="auth-button signin-btn"
-            whileHover={{ scale: 1.05 }}
-            whileTap={{ scale: 0.95 }}
-            initial={{ opacity: 0, x: 20 }}
-            animate={{ opacity: 1, x: 0 }}
-            transition={{ delay: 0.5 }}
-          >
-            Sign In
-          </motion.button>
-          <motion.Link
-            to="/auth/register"
-            className="auth-button signup-btn"
-            whileHover={{
-              scale: 1.05,
-              boxShadow: "0 8px 25px rgba(74, 98, 138, 0.3)",
-            }}
-            whileTap={{ scale: 0.95 }}
-            initial={{ opacity: 0, x: 20 }}
-            animate={{ opacity: 1, x: 0 }}
-            transition={{ delay: 0.6 }}
-          >
-            Get Started
-          </motion.Link>
-        </div>
-
-        {/* Mobile Menu Toggle */}
-        <motion.button
-          className="mobile-menu-toggle"
-          onClick={() => setIsMobileMenuOpen(!isMobileMenuOpen)}
-          whileTap={{ scale: 0.95 }}
-          initial={{ opacity: 0 }}
-          animate={{ opacity: 1 }}
-          transition={{ delay: 0.7 }}
-        >
-          <span className={`hamburger ${isMobileMenuOpen ? "active" : ""}`}>
-            <span></span>
-            <span></span>
-            <span></span>
+              <span className="nav-link__icon material-symbols-outlined-fill">
+                {" "}
+                {link.icon}
+              </span>
+              <span className="nav-link__text">{link.title}</span>
+              {link.sub && <input type="checkbox" id={index} name="nav-link" />}
+            </ContainerType>
           </span>
-        </motion.button>
-      </div>
 
-      {/* Mobile Menu */}
-      <AnimatePresence>
-        {isMobileMenuOpen && (
-          <motion.div
-            initial={{ opacity: 0, height: 0 }}
-            animate={{ opacity: 1, height: "auto" }}
-            exit={{ opacity: 0, height: 0 }}
-            transition={{ duration: 0.3, ease: "easeInOut" }}
-            className="mobile-menu"
-          >
-            <div className="mobile-menu-content">
-              {navLinks.map((link, index) => (
-                <motion.a
-                  key={link.name}
-                  href={link.href}
-                  onClick={(e) => {
-                    e.preventDefault();
-                    handleNavClick(link.href);
-                  }}
-                  className="mobile-nav-link"
-                  initial={{ opacity: 0, x: -20 }}
-                  animate={{ opacity: 1, x: 0 }}
-                  transition={{ delay: index * 0.1 }}
-                >
-                  {link.name}
-                </motion.a>
-              ))}
-              <div className="mobile-auth-buttons">
-                <button
-                  onClick={() => {
-                    onSignIn();
-                    setIsMobileMenuOpen(false);
-                  }}
-                  className="mobile-auth-button signin"
-                >
-                  Sign In
-                </button>
+          <ul className="nav-links__sub">
+            {link.sub?.map((subLink, subIndex) => (
+              <li
+                key={`index-sub-li-${index}-${subIndex}`}
+                className={`nav-link ${
+                  isNavLinkActive(subLink) ? "selected" : ""
+                }`}
+                title={subLink.title}
+              >
                 <Link
-                  to="/auth/register"
-                  className="mobile-auth-button signup"
-                  onClick={() => setIsMobileMenuOpen(false)}
+                  className="link"
+                  to={subLink.path[0]}
+                  key={`index-sub_${subIndex}`}
                 >
-                  Get Started
+                  <span className="nav-link__icon material-symbols-outlined-fill">
+                    {" "}
+                    {subLink.icon}
+                  </span>
+                  <span className="nav-link__text">{subLink.title}</span>
                 </Link>
-              </div>
-            </div>
-          </motion.div>
-        )}
-      </AnimatePresence>
-    </motion.nav>
-  );
-};
+              </li>
+            ))}
+          </ul>
+        </label>
+      </li>
+    );
+  });
 
-export default Navbar;
+  return (
+    <div className="nav">
+      <div className="nav-container">
+        <div className="nav__logo-container">
+          <div className="nav__logo">{icon}</div>
+        </div>
+        <div className="nav__name">
+          <div className="nav__name-details">Auto Connect</div>
+          <div className="nav__name-version">{version}</div>
+        </div>
+      </div>
+      <div className="nav__wrapper">
+        <ul className="nav-links">{navLinkElements}</ul>
+      </div>
+      <div className="nav-container">
+        <div className="nav__profile-image-container">
+          <div className="nav__profile-image"></div>
+        </div>
+        <div className="nav__profile">
+          <div className="nav__profile-details">John Doe</div>
+          <Link to="/auth/logout">
+            <div className="nav__logout material-symbols-outlined-fill">
+              logout
+            </div>
+          </Link>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default NavBar;
