+++ auto-connect-frontend/src/components/TopNavBar.jsx
@@ -1,0 +1,189 @@
+
+import React, { useState, useEffect } from "react";
+import { Link } from "react-router-dom";
+import { motion, AnimatePresence } from "framer-motion";
+import "./Navbar.css";
+
+const Navbar = ({ onSignIn }) => {
+  const [isScrolled, setIsScrolled] = useState(false);
+  const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
+
+  useEffect(() => {
+    const handleScroll = () => {
+      setIsScrolled(window.scrollY > 50);
+    };
+
+    window.addEventListener("scroll", handleScroll);
+    return () => window.removeEventListener("scroll", handleScroll);
+  }, []);
+
+  const navLinks = [
+    { name: "Features", href: "#features" },
+    { name: "How It Works", href: "#how-it-works" },
+    { name: "Testimonials", href: "#testimonials" },
+    { name: "Contact", href: "#contact" },
+  ];
+
+  const handleNavClick = (href) => {
+    if (href.startsWith("#")) {
+      const element = document.querySelector(href);
+      if (element) {
+        element.scrollIntoView({ behavior: "smooth" });
+      }
+    }
+    setIsMobileMenuOpen(false);
+  };
+
+  return (
+    <motion.nav
+      initial={{ y: -100 }}
+      animate={{ y: 0 }}
+      transition={{ duration: 0.6, ease: "easeOut" }}
+      className={`navbar ${isScrolled ? "scrolled" : ""}`}
+    >
+      <div className="nav-container">
+        {/* Logo */}
+        <motion.div whileHover={{ scale: 1.05 }} className="nav-logo">
+          <Link to="/" className="logo-link">
+            <div className="logo-icon">
+              <svg
+                viewBox="0 0 24 24"
+                fill="none"
+                xmlns="http://www.w3.org/2000/svg"
+              >
+                <path
+                  d="M3 18V16C3 14.8954 3.89543 14 5 14H19C20.1046 14 21 14.8954 21 16V18M3 18V20C3 21.1046 3.89543 22 5 22H19C20.1046 22 21 21.1046 21 20V18M3 18H21M12 2L8 6H16L12 2Z"
+                  stroke="currentColor"
+                  strokeWidth="2"
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                />
+                <circle cx="7" cy="18" r="1" fill="currentColor" />
+                <circle cx="17" cy="18" r="1" fill="currentColor" />
+              </svg>
+            </div>
+            <span className="logo-text">AutoConnect</span>
+          </Link>
+        </motion.div>
+
+        {/* Desktop Navigation */}
+        <div className="nav-links desktop-nav">
+          {navLinks.map((link, index) => (
+            <motion.a
+              key={link.name}
+              href={link.href}
+              onClick={(e) => {
+                e.preventDefault();
+                handleNavClick(link.href);
+              }}
+              className="nav-link"
+              whileHover={{ y: -2 }}
+              initial={{ opacity: 0, y: -20 }}
+              animate={{ opacity: 1, y: 0 }}
+              transition={{ delay: index * 0.1 + 0.3 }}
+            >
+              {link.name}
+            </motion.a>
+          ))}
+        </div>
+
+        {/* Auth Buttons */}
+        <div className="nav-auth desktop-nav">
+          <motion.button
+            onClick={onSignIn}
+            className="auth-button signin-btn"
+            whileHover={{ scale: 1.05 }}
+            whileTap={{ scale: 0.95 }}
+            initial={{ opacity: 0, x: 20 }}
+            animate={{ opacity: 1, x: 0 }}
+            transition={{ delay: 0.5 }}
+          >
+            Sign In
+          </motion.button>
+          <motion.Link
+            to="/auth/register"
+            className="auth-button signup-btn"
+            whileHover={{
+              scale: 1.05,
+              boxShadow: "0 8px 25px rgba(74, 98, 138, 0.3)",
+            }}
+            whileTap={{ scale: 0.95 }}
+            initial={{ opacity: 0, x: 20 }}
+            animate={{ opacity: 1, x: 0 }}
+            transition={{ delay: 0.6 }}
+          >
+            Get Started
+          </motion.Link>
+        </div>
+
+        {/* Mobile Menu Toggle */}
+        <motion.button
+          className="mobile-menu-toggle"
+          onClick={() => setIsMobileMenuOpen(!isMobileMenuOpen)}
+          whileTap={{ scale: 0.95 }}
+          initial={{ opacity: 0 }}
+          animate={{ opacity: 1 }}
+          transition={{ delay: 0.7 }}
+        >
+          <span className={`hamburger ${isMobileMenuOpen ? "active" : ""}`}>
+            <span></span>
+            <span></span>
+            <span></span>
+          </span>
+        </motion.button>
+      </div>
+
+      {/* Mobile Menu */}
+      <AnimatePresence>
+        {isMobileMenuOpen && (
+          <motion.div
+            initial={{ opacity: 0, height: 0 }}
+            animate={{ opacity: 1, height: "auto" }}
+            exit={{ opacity: 0, height: 0 }}
+            transition={{ duration: 0.3, ease: "easeInOut" }}
+            className="mobile-menu"
+          >
+            <div className="mobile-menu-content">
+              {navLinks.map((link, index) => (
+                <motion.a
+                  key={link.name}
+                  href={link.href}
+                  onClick={(e) => {
+                    e.preventDefault();
+                    handleNavClick(link.href);
+                  }}
+                  className="mobile-nav-link"
+                  initial={{ opacity: 0, x: -20 }}
+                  animate={{ opacity: 1, x: 0 }}
+                  transition={{ delay: index * 0.1 }}
+                >
+                  {link.name}
+                </motion.a>
+              ))}
+              <div className="mobile-auth-buttons">
+                <button
+                  onClick={() => {
+                    onSignIn();
+                    setIsMobileMenuOpen(false);
+                  }}
+                  className="mobile-auth-button signin"
+                >
+                  Sign In
+                </button>
+                <Link
+                  to="/auth/register"
+                  className="mobile-auth-button signup"
+                  onClick={() => setIsMobileMenuOpen(false)}
+                >
+                  Get Started
+                </Link>
+              </div>
+            </div>
+          </motion.div>
+        )}
+      </AnimatePresence>
+    </motion.nav>
+  );
+};
+
+export default Navbar;
