+++ auto-connect-frontend/src/css/tailwind.css
@@ -1,97 +1,2 @@
-@import "tailwindcss";
-
-/* Custom styles that work with Tailwind CSS v4 and your color scheme */
-:root {
-  --primary-light: #DFF2EB;
-  --primary-light-blue: #B9E5E8;
-  --primary-medium: #7AB2D3;
-  --primary-dark: #4A628A;
-}
-
-@layer base {
-  body {
-    font-feature-settings: "rlig" 1, "calt" 1;
-    font-family: "Poppins", "Roboto", "work sans", sans-serif;
-  }
-}
-
-@layer components {
-  .btn {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    border-radius: 0.375rem;
-    font-size: 0.875rem;
-    font-weight: 500;
-    transition-property: color, background-color, border-color, text-decoration-color, fill, stroke;
-    transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
-    transition-duration: 150ms;
-  }
-  
-  .btn:focus-visible {
-    outline: 2px solid transparent;
-    outline-offset: 2px;
-    box-shadow: 0 0 0 2px var(--primary-medium);
-  }
-  
-  .btn:disabled {
-    pointer-events: none;
-    opacity: 0.5;
-  }
-  
-  .btn-primary {
-    background-color: var(--primary-medium);
-    color: white;
-    box-shadow: 0 1px 2px 0 rgb(0 0 0 / 0.05);
-  }
-  
-  .btn-primary:hover {
-    opacity: 0.9;
-  }
-  
-  .btn-secondary {
-    background-color: var(--primary-light-blue);
-    color: var(--primary-dark);
-    box-shadow: 0 1px 2px 0 rgb(0 0 0 / 0.05);
-  }
-  
-  .btn-secondary:hover {
-    opacity: 0.8;
-  }
-
-  .card-primary {
-    background-color: var(--primary-light);
-    border: 1px solid var(--primary-light-blue);
-    color: var(--primary-dark);
-  }
-}
-
-@layer utilities {
-  .text-balance {
-    text-wrap: balance;
-  }
-  
-  .bg-primary-light {
-    background-color: var(--primary-light);
-  }
-  
-  .bg-primary-light-blue {
-    background-color: var(--primary-light-blue);
-  }
-  
-  .bg-primary-medium {
-    background-color: var(--primary-medium);
-  }
-  
-  .bg-primary-dark {
-    background-color: var(--primary-dark);
-  }
-  
-  .text-primary-dark {
-    color: var(--primary-dark);
-  }
-  
-  .text-primary-medium {
-    color: var(--primary-medium);
-  }
-}
+/* stylelint-disable-next-line */
+@import "tailwindcss" prefix(tw);
